Restructuring of packages: deleted package "test" and merged it's content with existing packages (see TestController). Now the list of tests can be accessed via /student/test/{test_id}. Fixed bug in tests which occured after sending the answers to server (the newly requested web-page had no content).
