+++ src/main/java/com/ncedu/testing/entity/Question.java
@@ -1,6 +1,7 @@
 package com.ncedu.testing.entity;
 
 import javax.persistence.*;
+import java.util.ArrayList;
 
 @Entity
 @Table(name = "questions")
@@ -20,14 +21,21 @@ public class Question {
     private String cAnswer;
 
     @ManyToOne(optional=false)
-    @JoinColumn(name="TEST_ID",referencedColumnName = "TEST_ID")
+    @JoinColumn(name = "TEST_ID",referencedColumnName = "TEST_ID")
     private Test test;
+
+    private ArrayList<String> answers = new ArrayList<>();
 
     public Question(String text, String options, String cAnswer, Test test) {
         this.text = text;
         this.options = options;
         this.cAnswer = cAnswer;
         this.test = test;
+        this.answers = new ArrayList<String>();
+        String[] answ = options.split("-");
+        for(String ans : answ){
+            this.answers.add(ans);
+        }
     }
 
     public Question() {
@@ -52,6 +60,13 @@ public class Question {
         result = 31 * result + options.hashCode();
         result = 31 * result + test.hashCode();
         return result;
+    }
+
+    public void genAnswers() {
+        String[] answ = options.split("-");
+        for(String ans : answ){
+            answers.add(ans);
+        }
     }
 
     public Long getId() {
@@ -92,5 +107,13 @@ public class Question {
 
     public void setText(String text) {
         this.text = text;
+    }
+
+    public ArrayList<String> getAnswers() {
+        return answers;
+    }
+
+    public void setAnswers(ArrayList<String> answers) {
+        this.answers = answers;
     }
 }
